completely moved domain to own module
