+++ app/src/main/java/luci/sixsixsix/powerampache2/data/SettingsRepositoryImpl.kt
@@ -31,8 +31,8 @@ import luci.sixsixsix.powerampache2.data.local.entities.toLocalSettingsEntity
 import luci.sixsixsix.powerampache2.data.remote.MainNetwork
 import luci.sixsixsix.powerampache2.domain.SettingsRepository
 import luci.sixsixsix.powerampache2.domain.errors.ErrorHandler
-import luci.sixsixsix.powerampache2.domain.models.LocalSettings
-import luci.sixsixsix.powerampache2.domain.models.SortMode
+import luci.sixsixsix.powerampache2.domain.models.settings.LocalSettings
+import luci.sixsixsix.powerampache2.domain.models.settings.SortMode
 import luci.sixsixsix.powerampache2.domain.utils.StorageManager
 import javax.inject.Inject
 import javax.inject.Singleton
