+++ app/src/main/java/luci/sixsixsix/powerampache2/di/RepositoryModule.kt
@@ -25,7 +25,6 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import luci.sixsixsix.powerampache2.common.ConfigProviderImpl
 import luci.sixsixsix.powerampache2.data.AlbumsRepositoryImpl
 import luci.sixsixsix.powerampache2.data.AmpachePreferencesRepositoryImpl
 import luci.sixsixsix.powerampache2.data.ArtistsRepositoryImpl
@@ -44,7 +43,6 @@ import luci.sixsixsix.powerampache2.domain.MusicRepository
 import luci.sixsixsix.powerampache2.domain.PlaylistsRepository
 import luci.sixsixsix.powerampache2.domain.SettingsRepository
 import luci.sixsixsix.powerampache2.domain.SongsRepository
-import luci.sixsixsix.powerampache2.domain.utils.ConfigProvider
 import luci.sixsixsix.powerampache2.domain.utils.ShareManager
 import luci.sixsixsix.powerampache2.domain.utils.SharedPreferencesManager
 import luci.sixsixsix.powerampache2.domain.utils.StorageManager
