+++ domain/src/main/java/luci/sixsixsix/powerampache2/domain/models/settings/SortMode.kt
@@ -1,0 +1,33 @@
+/**
+ * Copyright (C) 2025  Antonio Tari
+ *
+ * This file is a part of Power Ampache 2
+ * Ampache Android client application
+ * @author Antonio Tari
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+package luci.sixsixsix.powerampache2.domain.models.settings
+
+import luci.sixsixsix.powerampache2.domain.models.settings.LocalSettings.Companion.SETTINGS_DEFAULTS_PLAYLIST_SORT
+
+const val SORT_MODE_ASC = "ASC"
+private const val SORT_MODE_DESC = "DESC"
+val defaultPlaylistSort = SortMode.valueOf(SETTINGS_DEFAULTS_PLAYLIST_SORT)
+
+enum class SortMode(mode: String) {
+    ASC(SORT_MODE_ASC),
+    DESC(SORT_MODE_DESC)
+}
